+++ lib/appium_lib_core/common/base/driver.rb
@@ -21,6 +21,7 @@ require_relative 'rotable'
 require_relative 'remote_status'
 require_relative 'has_location'
 require_relative 'has_network_connection'
+require_relative '../wait'
 
 module Appium
   module Core
@@ -37,6 +38,8 @@ module Appium
         include ::Appium::Core::Base::HasRemoteStatus
         include ::Appium::Core::Base::HasLocation
         include ::Appium::Core::Base::HasNetworkConnection
+
+        include ::Appium::Core::Waitable
 
         private